refactor: Divide file to multiple packages
