+++ ids/ids.go
@@ -1,0 +1,16 @@
+package ids
+
+// Action IDs
+const (
+	SelectMenuByUser = "select_menu_by_user"
+	SubmitMenu       = "submit_menu"
+	AddMenuButton    = "add_menu"
+	TerminateMenu    = "terminate_menu"
+)
+
+// Block IDs
+const (
+	SubmitMenuBlock        = "submit_menu_block"
+	MenuButtonsBlock       = "menu_buttons_block"
+	MenuSelectContextBlock = "menu_select_context_block/"
+)
